Moving configuration values and methods from install.py to config.py -- everything is broken right now
